Add docroot and public_html symlinks to web as a sane default.
"web" folder is where Drupal resides. Is a standard name used by other
boilerplates, and a way to find Drupal root easily.

"docroot" is the Drupal docroot (you can say the Drupal root directory).

"public_html" is the root directory that the web server serves.

Using this different names ("web"·, "public_html" and "docroot") you can
configure the project to adapt to different setups. For example, the
most common setup, a Drupal isnbtalled in the web server root folder,
"docroot" and "public_html" point to "web" folder. But, for example if
Drupal is installed in a webserver subdirectory, say "drupal" folder,
then public_html will be a directory (instead of a symlink), and inside
it will have a symlink "drupal" that points to the "web" directory.